fix: reverts back socket.io code and increases rate limits
