+++ packages/sdk-socket-server/.env.sample
@@ -1,3 +1,5 @@
 SEGMENT_API_KEY_PRODUCTION=123456789
 SEGMENT_API_KEY_DEBUG=123456789
 REDIS_URL=redis://redis:6379/0
+RATE_LIMIT_POINTS_MAX=200
+RATE_LIMIT_MESSAGE_POINTS_MAX=2000
