+++ packages/sdk-socket-server/Dockerfile
@@ -1,27 +1,15 @@
-# Build stage
-FROM node:18-alpine AS builder
-
-# Install build dependencies and build the project
-WORKDIR /app
-COPY package.json ./
-RUN yarn install
-COPY . .
-RUN yarn build
-
-# Runtime stage
 FROM node:18-alpine
 
-# Install runtime dependencies
+RUN apk update && apk upgrade && apk add --no-cache zlib
+
 WORKDIR /app
-COPY --from=builder /app/package.json ./
-RUN yarn install --production
 
-# Copy built project and .env file from the build stage
-COPY --from=builder /app/dist ./dist
-COPY .env ./
+COPY package.json .
 
-# Expose the server port
+RUN yarn install
+
+COPY . .
+
 EXPOSE 4000
 
-# Start the server
-CMD ["node", "dist/index.js"]
+CMD ["yarn","start"]
