--- packages/sdk-socket-server/nodemon.json
@@ -1,5 +1,0 @@
-{
-  "watch": ["."],
-  "ext": "ts",
-  "exec": "ts-node"
-}
