--- packages/sdk-socket-server/tsconfig.test.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES6",
-    "module": "commonjs",
-    "outDir": "dist",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "types": ["node", "jest", "@testing-library/jest-dom"]
-  },
-  "include": ["*.ts"],
-  "exclude": ["node_modules"]
-}
